create component for api and quotes
